address review comment and update get_daily_question endpoint
