+++ app/api/routes/questions.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter
-from app.services.questions import generate_daily_question as generate_daily_question_service, add_daily_question_to_store as add_question_in_chroma_db, add_daily_question_to_mongodb
+from app.services.questions import generate_daily_question as generate_daily_question_service, add_daily_question_to_store as add_question_in_chroma_db, add_daily_question_to_mongodb, get_most_recent_daily_question
 from app.schemas.utils import UserPrompt
+
 router = APIRouter()
 
 
@@ -19,6 +20,7 @@ def get_daily_question():
     """
     Retrieve the most recent daily question.
     """
-    #TODO: implement this endpoint
-    # Get from mongoDB
-    return {"message": "This endpoint will return the most recent daily question. Not implemented yet."}
+    recent_question = get_most_recent_daily_question()
+    if recent_question:
+        return recent_question
+    return {"message": "No daily question found."}
