--- config/database.py
@@ -1,13 +1,0 @@
-from pymongo import MongoClient
-
-class MongoDB:
-    def __init__(self, uri: str, db_name: str):
-        self.client = MongoClient(uri)
-        self.db = self.client[db_name]
-
-    def get_collection(self, collection_name: str):
-        return self.db[collection_name]
-
-# Example usage
-mongo = MongoDB(uri="mongodb://mongodb:27017", db_name="questions_db")
-questions_collection = mongo.get_collection("questions")
